+++ resources/views/components/layout.blade.php
@@ -10,17 +10,25 @@
 </head>
 <body>
     <nav class="navbar bg-dark border-bottom border-body">
-        <x-navlink href="/" :active="request()->is('home')">Home</x-navlink>
+        <x-navlink href="/" :active="request()->is('home')"><img src="/images/logosmall.png"></x-navlink>
         <x-navlink href="{{ route('articles.index') }}" :active="request()->is('articles.index')">Articles</x-navlink>
-        <x-navlink href="{{ route('articles.create') }}" :active="request()->is('articles.index')">Post</x-navlink>
-        <x-navlink href="/dashboard" :active="request()->is('dashboard')">Dashboard</x-navlink>
-        <x-navlink href="/login" :active="request()->is('log in')">Log In</x-navlink>
+        @auth()
+            @if(Auth::user()->is_admin)
+                <x-nav-link href="{{route('admin.index')}}">Admin Dashboard</x-nav-link>
+            @endif
+            <x-navlink href="{{ route('articles.create') }}" :active="request()->is('articles.index')">Write article</x-navlink>
+            <x-navlink href="{{ route('dashboard') }}" :active="request()->is('dashboard')">{{  Auth::user()->name }}</x-navlink>
+        @endauth
+        @guest
+            <x-navlink href="{{ route('login') }}">Log In</x-navlink>
+            <x-nav-link href="{{ route('register') }}">Register</x-nav-link>
+        @endguest
     </nav>
     <main>
         {{ $slot }}
     </main>
     <footer>
-{{--        <x-navlink href="/logout">Log Out</x-navlink>--}}
+        <img src="/images/logo.png">
         <p>Made by Alex Dusselaar - 2024</p>
     </footer>
 </body>