+++ resources/views/home.blade.php
@@ -1,6 +1,14 @@
 <x-layout title="Home">
-    <h1>Welcome to Cosplay and Craft Community!</h1>
-    <section id="home">
-        <p>The forum for all creative souls looking for tips and tricks for all kinds of crafts!</p>
+    <header>
+        <img src="/images/logowhite.png">
+        <h5>The forum for all cosplay and craft tips, tricks and advice!</h5>
+    </header>
+     <section id="home">
+        <h2>What is cosplay?</h2>
+        <div>
+            <p>Cosplay, short for "costume play," is a performance art in which participants, known as cosplayers, dress up in costumes and accessories to represent a specific character from various media, including anime, video games, movies, and comic books. This creative practice often involves not only the visual aspects of costume creation but also embodying the character's personality and mannerisms, allowing cosplayers to immerse themselves in the role. Cosplay has grown into a vibrant subculture, featuring conventions, competitions, and online communities where enthusiasts share their craft, connect with others, and celebrate their passion for fandom.</p>
+            <img src="https://i.pinimg.com/564x/6c/e2/c9/6ce2c9c5e40bcf87e797ef269cec432e.jpg">
+        </div>
+        <p>Making your own cosplays is a fulfilling and creative endeavor that allows fans to bring their favorite characters to life in a personal way. The process typically begins with research, gathering reference images, and selecting materials that match the character's design. From sewing garments to crafting props, each step involves a combination of artistic expression and technical skills, such as pattern-making, dyeing fabrics, and using tools like sewing machines or hot glue guns. Many cosplayers enjoy experimenting with techniques, whether it’s using foam to create armor or learning how to style wigs for authenticity. The journey of creating a cosplay is not just about the final product; it’s also a chance to learn, connect with fellow enthusiasts, and gain confidence in one’s craftsmanship. Plus, wearing a self-made costume at conventions or events adds a unique sense of pride and accomplishment, making the experience even more rewarding.</p>
     </section>
 </x-layout>